SchemaDomain column types, decimal columns BigDecimal instead of Double, floating point columns no longer Integer, whole numbers beyond an Integer Long, multi character CHAR columns String, NCHAR, NVARCHAR, CLOB, NCLOB, TINYINT and binary columns now mapped, time zone, UUID and Oracle binary float columns recognized by type name, SQLite columns based on their declared types SchemaDomain, catalogs used as schemas when a database reports no schemas, MySQL and MariaDB, which previously required the databaseTerm=SCHEMA or useCatalogTerm=Schema driver property DomainSource no longer outputs fractionDigits() when equal to the default

+++ framework/domain-db/src/main/java/is/codion/framework/domain/db/MetaDataColumn.java
@@ -20,6 +20,7 @@ package is.codion.framework.domain.db;
 
 import is.codion.common.db.result.ResultPacker;
 
+import java.math.BigDecimal;
 import java.sql.JDBCType;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -241,13 +242,14 @@ final class MetaDataColumn {
 		}
 	}
 
-	// A whole number is an Integer or a Long depending on its precision, which may exceed that of a Long
+	// A whole number is an Integer or a Long depending on its precision, which may exceed that of a Long,
+	// a decimal number, or one of unknown scale, a BigDecimal, the exact type
 	private static Class<?> numericType(int precision, int scale) {
 		if (scale == 0 && precision > 0) {
 			return precision <= 9 ? Integer.class : Long.class;
 		}
 
-		return Double.class;
+		return BigDecimal.class;
 	}
 
 	private static int declaredType(String typeName, int sqlType) {
